projects/feature-engineering-dimred-showcase/pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "feature-engineering-dimred-showcase"
version = "0.1.0"
description = "Feature engineering and dimensionality reduction showcase."
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
  "matplotlib>=3.9.0",
  "numpy>=1.26.0",
  "pandas>=2.2.0",
  "scikit-learn>=1.6.0",
]

[project.optional-dependencies]
advanced = [
  "autofeat>=2.1.3",
  "featuretools>=1.30.0",
  "tsfresh>=0.21.0",
  "umap-learn>=0.5.6",
]
dev = [
  "mypy>=1.11.2",
  "pytest>=8.3.3",
  "ruff>=0.9.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/feature_dimred_showcase"]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
mypy_path = "src"
